+++ src/terminal/sys.rs
@@ -4,12 +4,18 @@
 #[cfg(feature = "events")]
 pub use self::unix::supports_keyboard_enhancement;
 #[cfg(unix)]
+#[cfg(feature = "events")]
+pub use self::unix::title;
+#[cfg(unix)]
 pub(crate) use self::unix::{
     disable_raw_mode, enable_raw_mode, is_raw_mode_enabled, size, window_size,
 };
 #[cfg(windows)]
 #[cfg(feature = "events")]
 pub use self::windows::supports_keyboard_enhancement;
+#[cfg(windows)]
+#[cfg(feature = "events")]
+pub use self::windows::title;
 #[cfg(all(windows, test))]
 pub(crate) use self::windows::temp_screen_buffer;
 #[cfg(windows)]